--- a/geos/libgeos.go
+++ b/geos/libgeos.go
@@ -507,10 +507,10 @@ func (h *handle) decode(gh *C.GEOSGeometry, opts []geom.ConstructorOption) (geom
 		return geom.Geometry{}, h.err()
 	}
 	defer C.GEOSFree_r(h.context, unsafe.Pointer(serialised))
-	r := bytes.NewReader(C.GoBytes(unsafe.Pointer(serialised), C.int(size)))
+	byts := C.GoBytes(unsafe.Pointer(serialised), C.int(size))
 
 	if isWKT != 0 {
-		return geom.UnmarshalWKT(r, opts...)
+		return geom.UnmarshalWKT(bytes.NewReader(byts), opts...)
 	}
-	return geom.UnmarshalWKB(r, opts...)
+	return geom.UnmarshalWKB(byts, opts...)
 }